Fix LinkageError caused by invoke
The invoke namespace could result in LinkageErrors being caused in cases
where multiple Clojure runtimes are used in the same JVM (e.g.
shimdandy). Since invoke was using a MethodHandle lookup in the context
of Object, a LinkageError could result because the Clojure classes would
be loaded from different classloaders. By putting the lookup in the
context of clojure.lang.RT, it should always be scoped to the Clojure
runtime's classloader, instead of the JVM's platform/extensions one.