Change stream test to work on Java 9
Clojure is generating the wrong bytecode on Java 9+ for static methods
on interfaces. Since, IntStream is an interface the IntStream/range
calls were failing due to this. Switched to an alternante way to get a
stream so the tests will work.

See Clojure bug: https://dev.clojure.org/jira/browse/CLJ-2284